[PR] 배포 테스트

+++ src/main/java/com/hongik/mentor/hongik_mentor/service/PostService.java
@@ -3,17 +3,19 @@ package com.hongik.mentor.hongik_mentor.service;
 import com.hongik.mentor.hongik_mentor.controller.dto.PostCreateDTO;
 import com.hongik.mentor.hongik_mentor.controller.dto.PostDTO;
 import com.hongik.mentor.hongik_mentor.controller.dto.PostModifyDTO;
+import com.hongik.mentor.hongik_mentor.controller.dto.comment.CommentModifyDto;
+import com.hongik.mentor.hongik_mentor.controller.dto.comment.CommentReqDto;
+import com.hongik.mentor.hongik_mentor.controller.dto.comment.CommentResDto;
 import com.hongik.mentor.hongik_mentor.domain.*;
-import com.hongik.mentor.hongik_mentor.domain.post.Post;
-import com.hongik.mentor.hongik_mentor.domain.post.PostLike;
-import com.hongik.mentor.hongik_mentor.domain.post.PostTag;
-import com.hongik.mentor.hongik_mentor.domain.post.Tag;
+import com.hongik.mentor.hongik_mentor.domain.member.Member;
+import com.hongik.mentor.hongik_mentor.domain.post.*;
 import com.hongik.mentor.hongik_mentor.exception.CustomMentorException;
 import com.hongik.mentor.hongik_mentor.exception.ErrorCode;
 import com.hongik.mentor.hongik_mentor.repository.MemberRepository;
 import com.hongik.mentor.hongik_mentor.repository.PostRepository;
 import com.hongik.mentor.hongik_mentor.repository.TagRepository;
 import jakarta.persistence.OptimisticLockException;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -32,36 +34,47 @@ public class PostService {
     private final TagRepository tagRepository;
 
     @Transactional
-    public Long createPost(PostCreateDTO postCreateDTO) {
+    public Long createPost(PostCreateDTO postCreateDTO, Long memberId) {
 
-        Member member = memberRepository.findById(postCreateDTO.getMemberId())
-                .orElseThrow(() -> new CustomMentorException(ErrorCode.MEMBER_NOT_EXISTS));
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(()->new CustomMentorException(ErrorCode.MEMBER_NOT_EXISTS));
 
         Post post = Post.builder()
                 .member(member)
                 .title(postCreateDTO.getTitle())
                 .content(postCreateDTO.getContent())
+                .category(postCreateDTO.getCategory())
+                .capacity(postCreateDTO.getCategory() == Category.MENTOR ? postCreateDTO.getCapacity() : 1)
                 .build();
 
-
-        postCreateDTO.getTagId()
-                .forEach(id -> {
-                    Tag tag = tagRepository.findById(id).orElseThrow(() -> new RuntimeException("Tag not found"));
-                    PostTag postTag = PostTag.of(tag, post);
-                    post.addTags(postTag);
-                });
+        if(!postCreateDTO.getTagIds().isEmpty()){
+            postCreateDTO.getTagIds()
+                    .forEach(id -> {
+                        Tag tag = tagRepository.findById(id).orElseThrow(() -> new RuntimeException("Tag not found"));
+                        PostTag postTag = PostTag.of(tag, post);
+                        post.addTags(postTag);
+                    });
+        }
 
         postRepository.save(post);
 
         return post.getId();
     }
 
-    public PostDTO getPost(Long postId){
+    public PostDTO getPost(Long postId, Long requesterId){    //추후 함수 getPost함수 이름을 별도로 빼서 순수하게 post조회하는 함수 생성해야할듯
 
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new CustomMentorException(ErrorCode.POST_NOT_EXISTS));
 
-        return PostDTO.fromPost(post);
+        List<CommentResDto> commentResDtos = post.getComments().stream()
+                .map(comment -> {
+                    boolean isCommentOwner = requesterId.equals(comment.getMember().getId());
+                    return new CommentResDto(comment, isCommentOwner);
+                }).toList();
+
+        boolean isOwner = requesterId.equals(post.getMember().getId());
+
+        return PostDTO.fromPost(post,isOwner,commentResDtos);
     }
 
     @Transactional
@@ -76,11 +89,12 @@ public class PostService {
 
         List<PostTag> postTags = tags.stream().map(tag -> PostTag.builder()
                 .post(post)
-                .tag(tag).build()).toList();
+                .tag(tag)
+                .build()).toList();
 
         post.clearTags();
 
-        post.modifyPost(postModifyDTO.getTitle(), postModifyDTO.getContent(), postTags);
+        post.modifyPost(postModifyDTO.getTitle(), postModifyDTO.getContent(), postTags, postModifyDTO.getCapacity());
 
         postRepository.save(post);
 
@@ -89,21 +103,29 @@ public class PostService {
     }
 
     @Transactional
-    public Long deletePost(Long postId){
-        postRepository.deleteById(postId);
+    public Long deletePost(Long postId, Long requesterId){
+        Post findPost = postRepository.findById(postId).orElseThrow();
+        if (isEntityOwner(requesterId, findPost.getMember().getId())) {
+            postRepository.deleteById(postId);
+        } else {
+            throw new RuntimeException("해당 게시글의 소유자가 아닙니다");
+        }
 
         return postId;
     }
 
-    public List<PostDTO> searchPostsByTags(List<Long> tagIds){ // 태그들 기반 검색
-        List<Post> posts = postRepository.searchByTags(tagIds);
 
-        if (posts.isEmpty()) {
-            throw new RuntimeException("Post not found");
-        }
+
+
+    public List<PostDTO> searchPostsByTags(Category category, List<Long> tagIds) {
+
+        List<Post> posts = postRepository.searchByTagsAndCategory(tagIds, category);
+
+        if(posts.isEmpty()) return List.of();
 
         return posts.stream()
-                .map(PostDTO::fromPost).toList();
+                .map(PostDTO::fromPost)
+                .toList();
     }
 
     @Transactional
@@ -132,7 +154,7 @@ public class PostService {
 
     // 모집 지원 기능
     @Transactional
-    public void applyToPost(Long postId, Long memberId) {
+    public void applyToPost(Long postId, Long memberId, String nickname) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new CustomMentorException(ErrorCode.POST_NOT_EXISTS));
 
@@ -140,7 +162,7 @@ public class PostService {
                 .orElseThrow(() -> new CustomMentorException(ErrorCode.MEMBER_NOT_EXISTS));
 
         try {
-            post.addApplicant(member);
+            post.addApplicant(member, nickname);
         } catch (OptimisticLockException e) {
             throw new RuntimeException("다시 시도해 주세요.");
         }
@@ -176,5 +198,47 @@ public class PostService {
 
         post.cancelApplicant(member);
         postRepository.save(post);
+    }
+
+    /** Tag
+     *
+     * */
+    public List<Tag> getTags() {
+        return tagRepository.findAll();
+    }
+
+    /**
+     * Comment
+     */
+    @Transactional
+    public void createComment(CommentReqDto dto, Long memberId) {
+        Post findPost = postRepository.findById(dto.getPostId()).orElseThrow(() -> new IllegalStateException("댓글을 작성하신 게시글이 존재하지 않습니다"));
+        Member findMember = memberRepository.findById(memberId).orElseThrow();
+
+        findPost.getComments().add(Comment.builder()
+                .post(findPost)
+                .member(findMember)
+                .content(dto.getComment())
+                .build());
+
+        postRepository.save(findPost);
+    }
+
+    @Transactional
+    public void modifyComment(Long postId, Long commentId, CommentModifyDto dto, HttpSession httpSession) {
+
+        //미완. 추가 작업 필요
+
+        /*Post findPost = postRepository.findById(postId).orElseThrow();
+
+        if(isEntityOwner(SessionUtil.getCurrentMemberId(httpSession)),)*/
+    }
+
+    /**
+     * Utility
+     * */
+    //요청자가 해당 엔티티의 주인인지 검증
+    private static boolean isEntityOwner(Long requesterId, Long entityOwnerId) {
+        return requesterId.equals(entityOwnerId);
     }
 }
